Added tokio integration tests
